+++ lib/src/embedded_cli.c
@@ -260,7 +260,7 @@ EmbeddedCliConfig *embeddedCliDefaultConfig(void) {
     return &defaultConfig;
 }
 
-size_t embeddedCliRequiredSize(EmbeddedCliConfig *config) {
+uint16_t embeddedCliRequiredSize(EmbeddedCliConfig *config) {
     uint16_t bindingCount = config->maxBindingCount + cliInternalBindingCount;
     return sizeof(EmbeddedCli) + sizeof(EmbeddedCliImpl) +
            config->rxBufferSize * sizeof(char) +
